refactor(prisma): change env for prisma to root
Moved env that prisma looks for to root in light of production env vars. All env vars should be
moved to `.env` including the DATABASE_URL.

BREAKING CHANGE: You must create a `.env` file with the env var DATABASE_URL as specified in
`.env.example`. Do NOT use any other `.env.[other]` files. Prisma will not be able to run if the env
vars are not moved.