Implementation of superjson (#100)
* fix(org migration): change lat and long from DECIMAL to DOUBLE PRECISION

This change gets around the Prisma change where DECIMAL uses the Decimal.js object structure. DOUBLE
PRECISION reverts back to the Float type that uses TS number type.

BREAKING CHANGE: As a direct change to a migration, the database will need to be reset.

* feat(package.json): add superjson

Implement Superjson and its Babel configuration change. Superjson will allow the seamless
integration with Date objects when passing the field in NextJS initial data fetches such as
getServerSideProps.

* feat: remove workaround for Date objects with JSON parse then stringify

With the implementation of SuperJSON, we are able to remove the workaround where the Date object
could not be parsed when passed in from NextJS initial data fetching. Full Date objects are now
supported without needing to do JSON.parse(JSON.stringify(obj)).

* refactor(map.tsx): remove workaround for decimal.js in map

Removed casting of Number around lat and long to convert Decimal object to number. This is now
unnecessary because lat and long are now Float which are numbers